Added an inspect option.
This option is used to validate that the pulled source seems legit.
Users should always check to make sure everything seems in order.
If it doesn't them you should not use this tool because your network is
compromised.

This option is made for users who know how to read/debug shellscripts